% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\alias{interpolate.character}
\alias{interpolate.factor}
\alias{interpolate.integer}
\alias{interpolate.numeric}
\title{Interpolate}
\usage{
interpolate(x, ...)

\method{interpolate}{numeric}(x, ninterp = 4L, ...)

\method{interpolate}{integer}(x, ninterp = 4L, ...)

\method{interpolate}{factor}(x, ninterp = 4L, ...)

\method{interpolate}{character}(x, ninterp = 4L, ...)
}
\arguments{
\item{x}{A numeric or factor vector.}

\item{...}{Not used.}

\item{ninterp}{The number of points to interpolate between observations. It
should be an even number for sensible results on a factor/character vector.}
}
\description{
Interpolate a numeric or factor vector.
}

