% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualweight.R
\name{visualweight}
\alias{visualweight}
\title{Calculate the visual weight for a set of observations}
\usage{
visualweight(xc.cond, xc, sigma = NULL, distance = NULL, constant = NULL)
}
\arguments{
\item{xc.cond}{This is a dataframe describing the sections/conditions.}

\item{xc}{This is a dataframe with same names as \code{xc.cond} representing
observed data points.}

\item{sigma}{This is a threshold distance outside which observations will be
assigned visual weight zero. This is numeric (and should be > 0).}

\item{distance}{The type of distance measure to be used, currently just two
types of Minkowski distance: \code{"euclidean"} (default), and
\code{"maxnorm"}.}

\item{constant}{A constant to multiply by the number of categorical
mismatches, before adding to the Minkowski distance, to give a general
dissimilarity measure.}
}
\value{
A numeric vector or matrix, with values between 0 and 1. The visual
  weights for the observations in \code{xc} arranged in rows for each row
  in \code{xc.cond}.
}
\description{
Calculate the visual weight for a set of observations, based on their
distance from some sections/conditions
}
\details{
Visual weight is assigned to observations based on their
  distance from a section (treated as an observation here). The distance is
  calculated as Minkowski distance between the numeric elements for the
  observations whose categorical elements match, with the option to use a
  more general dissimilarity measure comprising Minkowski distance and a
  mismatch count.
}

