% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condtour.R
\name{condtour}
\alias{condtour}
\title{Conditional tour; a tour through sections in data space}
\usage{
condtour(data, model, path, response = NULL, S = NULL, C = NULL,
  sigma = NULL, distance = c("euclidean", "maxnorm"), cex.axis = NULL,
  cex.lab = NULL, tck = NULL, view3d = FALSE, conf = FALSE,
  select.colour = "blue", col = "black", pch = 1)
}
\arguments{
\item{data}{A dataframe.}

\item{model}{A fitted model object, or a list of such objects.}

\item{path}{A dataframe, describing the sections to take. Basically a
dataframe with its \code{colnames} being a subset of the
\code{colnames(data)}.}

\item{response}{Character name of response variable in \code{data}.}

\item{S}{Character name(s) of variables in \code{data} on which to take
sections.}

\item{C}{Character name(s) of variables in \code{data} on which to condition.}

\item{sigma}{Threshold distance. Observed data which are a distance greater
than \code{sigma} from the current section are not visible.}

\item{distance}{The type of distance measure to use, either
\code{"euclidean"} (default) or \code{"maxnorm"}.}

\item{cex.axis}{Scaling for axis text}

\item{cex.lab}{Scaling for labels}

\item{tck}{Tick size for axes.}

\item{view3d}{Logical; if \code{TRUE}, plots a three-dimensional regression
surface when possible.}

\item{conf}{Logical; if \code{TRUE}, plots confidence bounds or equivalent
when possible.}

\item{select.colour}{Colour to highlight current section.}

\item{col}{Colour for observed data points.}

\item{pch}{Plot symbols for observed data points.}
}
\value{
Produces a set of interactive plots. One device displays the current
  section. A second device shows the the current section in the space of the
  conditioning predictors given by \code{C}. A third device shows some simple
  diagnostic plots; one to show approximately how much data are visible on
  each section, and another to show what proportion of data are
  \emph{visited} by the tour.
}
\description{
Whereas \code{\link{ceplot}} allows the user to interactively
  choose sections to visualise, \code{condtour} allows the user to pre-select
  all sections to visualise, order them, and cycle through them one by one.
}
\examples{
\dontrun{

data(powerplant)
library(e1071)
model <- svm(PE ~ ., data = powerplant)
path <- makepath(powerplant[-5], 25)
condtour(data = powerplant, model = model, path = path$path, S = "AT")

data(wine)
wine$Class <- as.factor(wine$Class)
library(e1071)
model5 <- list(svm(Class ~ ., data = wine))
C <- setdiff(colnames(wine), c("Class", "Hue", "Flavanoids"))
path <- makepath(wine[, C], 50)
condtour(data = wine, model = model5, path = path$path, S = c("Hue",
  "Flavanoids"), sigma = 3)

}
}
\seealso{
\code{\link{ceplot}}, \code{\link{similarityweight}}
}

